included waveform_features module; removed tools; don't donwload fim at setup while macking a tarball with 'sdist' command
travis not fixed downloading fim
