+++ setup.py
@@ -1,13 +1,18 @@
 # -*- coding: utf-8 -*-
 
 import os
+import platform
+import struct
 import sys
 
 from setuptools import setup
 
-from tools.fim_manager import download_spade_fim
-
 python_version_major = sys.version_info.major
+
+if python_version_major == 2:
+    from urllib import urlretrieve
+else:
+    from urllib.request import urlretrieve
 
 with open(os.path.join(os.path.dirname(__file__),
                        "elephant", "VERSION")) as version_file:
@@ -22,7 +27,35 @@ for extra in ['extras', 'docs', 'tests']:
     with open('requirements-{0}.txt'.format(extra)) as fp:
         extras_require[extra] = fp.read()
 
-download_spade_fim()
+
+def download_spade_fim():
+    """
+    Downloads SPADE specific PyFIM binary file.
+    """
+    if platform.system() == "Windows":
+        fim_filename = "fim.pyd"
+    else:
+        # Linux
+        fim_filename = "fim.so"
+    spade_src_dir = os.path.join(os.path.dirname(__file__), "elephant",
+                                 "spade_src")
+    fim_lib_path = os.path.join(spade_src_dir, fim_filename)
+    if os.path.exists(fim_lib_path):
+        return
+
+    arch_bits = struct.calcsize("P") * 8
+    url_fim = "http://www.borgelt.net/bin{arch}/py{py_ver}/{filename}". \
+        format(arch=arch_bits, py_ver=python_version_major,
+               filename=fim_filename)
+    try:
+        urlretrieve(url_fim, filename=fim_lib_path)
+        print("Successfully downloaded fim lib to {}".format(fim_lib_path))
+    except Exception:
+        print("Unable to download {url} module.".format(url=url_fim))
+
+
+if len(sys.argv) > 1 and sys.argv[1].lower() != 'sdist':
+    download_spade_fim()
 
 setup(
     name="elephant",
